Only use BoostedYAML for loading
