+++ daisylib-plugin/src/main/resources/example.yml
@@ -1,0 +1,3 @@
+initial-key: false
+
+config-version: 2
